this is exactly why I gave up on CSV::Row
exactly why I gave up on CSV::Row and CSV::Table last time.  It assumes
that your headers are consistent, even going so far as to assume that
they are passed in to create each CSV::Row in the same order.  (what is
the point of having both keys and values in a set operation if they are
not regarded as important?)